+++ src/v_video.h
@@ -126,7 +126,7 @@ void V_PutBlock(int x, int y, int scrn, int width, int height, byte *src);
 
 void V_DrawHorizLine(int x, int y, int scrn, int width, byte color);
 
-void V_ShadeScreen(void);
+void V_ShadeScreen(const int targshade); // [Nugget] Parameterized
 
 // [FG] colored blood and gibs